requirements.txt
# Web Framework
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
waitress>=2.1.2

# HTTP Client
requests>=2.31.0
httpx>=0.25.0
gql[requests]>=3.4.0

# Scheduling
apscheduler>=3.10.0

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
cryptography>=41.0.0

# Logging
structlog>=23.2.0

# Web UI
bootstrap-flask>=2.3.0
wtforms>=3.1.0
